ChametProject_API/chametApi/requirements.txt
selenium==4.6.0
requests==2.24.0
pytest==7.1.3
PyYAML==5.3.1
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
asgiref==3.3.4
async-lru==1.0.2
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
click==8.0.4
colorama==0.4.3
configparser==5.0.0
cryptography==2.9.2
cx-Oracle==8.3.0
Django==2.2.9
elasticsearch5==5.5.6
execnet==1.7.1
Faker==4.1.1
filelock==3.0.12
Flask==2.0.3
Flask-Cors==3.0.10
HTMLParser==0.0.2
idna==2.10
install==1.3.4
itsdangerous==2.1.0
Jinja2==3.0.3
jsonpath==0.82
MarkupSafe==2.1.0
more-itertools==8.4.0
MouseInfo==0.1.3
Naked==0.1.31
packaging==20.4
Pillow==8.1.0
pluggy==0.13.1
py==1.9.0
pyaes==1.6.1
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.11.0
PyGetWindow==0.0.9
pyjson==1.3.0
pykeyboard==0.1.1
PyMouse==1.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
pytest-check==2.1.4